requirements.txt
flask
obsws-python
pyyaml
requests
tomli
